+++ students/HABTAMU/lesson01/assignment/inventory_management/management.py
@@ -1,26 +1,9 @@
 """ Launches the user interface for the inventory management system """
 import sys
-import market_prices
-import inventoryclass
-import furnitureclass
-import electricappliancesclass
-
-
-def main_menu(user_prompt=None):
-    """ menu switcher methond """
-    valid_prompts = {"1": add_new_item,
-                     "2": item_info,
-                     "q": exit_program}
-    options = list(valid_prompts.keys())
-
-    while user_prompt not in valid_prompts:
-        options_str = ("{}" + (", {}") * (len(options)-1)).format(*options)
-        print("Please choose from the following options ({options_str}):")
-        print("1. Add a new item to the inventory")
-        print("2. Get item information")
-        print("q. Quit")
-        user_prompt = input(">")
-    return valid_prompts.get(user_prompt)
+from inventory_management import market_prices
+from inventory_management.inventoryclass import Inventory
+from inventory_management.furnitureclass import Furniture
+from inventory_management.electricappliancesclass import ElectricAppliances
 
 
 def get_price(item_code):
@@ -29,10 +12,8 @@ def get_price(item_code):
     print("Get price")
 
 
-def add_new_item():
+def add_new_item(FULLINVENTORY):
     """ add_new_item """
-
-    global FULLINVENTORY
     item_code = input("Enter item code: ")
     item_description = input("Enter item description: ")
     item_rentalprice = input("Enter item rental price: ")
@@ -44,21 +25,22 @@ def add_new_item():
     if is_furniture.lower() == "y":
         item_material = input("Enter item material: ")
         item_size = input("Enter item size (S,M,L,XL): ")
-        new_item = furnitureclass.Furniture(
+        new_item = Furniture(
             item_code,
             item_description,
             item_price,
             item_rentalprice,
             item_material,
             item_size
-            )
+        )
     else:
 
-        is_electricappliance = input("Is this item an electric appliance? (Y/N): ")
+        is_electricappliance = input(
+            "Is this item an electric appliance? (Y/N): ")
         if is_electricappliance.lower() == "y":
             item_brand = input("Enter item brand: ")
             item_voltage = input("Enter item voltage: ")
-            new_item = electricappliancesclass.ElectricAppliances(
+            new_item = ElectricAppliances(
                 item_code,
                 item_description,
                 item_price,
@@ -67,7 +49,7 @@ def add_new_item():
                 item_voltage
             )
         else:
-            new_item = inventoryclass.Inventory(
+            new_item = Inventory(
                 item_code,
                 item_description,
                 item_price,
@@ -75,9 +57,10 @@ def add_new_item():
             )
     FULLINVENTORY[item_code] = new_item.returnas_dictionary()
     print("New inventory item added")
+    return FULLINVENTORY
 
 
-def item_info():
+def item_info(FULLINVENTORY):
     """ item_info """
 
     item_code = input("Enter item code: ")
@@ -87,17 +70,10 @@ def item_info():
             print("{}:{}".format(k, value))
     else:
         print("Item not found in inventory")
+    return FULLINVENTORY
 
 
 def exit_program():
     """ exit_program """
 
     sys.exit()
-
-
-if __name__ == '__main__':
-    FULLINVENTORY = {}
-    while True:
-        print(FULLINVENTORY)
-        main_menu()()
-        input("Press Enter to continue...........")
